+++ ibaqpy/ibaq/utils.py
@@ -120,52 +120,6 @@ def fill_samples(df, proteins):
 
     return df
 
-
-def impute_missing_values(
-    data: Optional[Union[pd.DataFrame, List[pd.DataFrame]]],
-    n_neighbors=5,
-    weights="uniform",
-    metric="nan_euclidean",
-    keep_empty_features=True,
-) -> Union[pd.DataFrame, List[pd.DataFrame]]:
-    """
-    Impute missing values in a DataFrame or each DataFrame in a list using KNNImputer.
-
-    Parameters
-    ----------
-    data : Union[pd.DataFrame, List[pd.DataFrame]]
-        A pandas DataFrame or list of pandas DataFrames with missing values.
-    n_neighbors : int, optional
-        Number of neighboring samples to use for imputation. Default is 5.
-    weights : str, optional
-        Weight function used in prediction. Default is 'uniform'.
-    metric : str, optional
-        Distance metric for searching neighbors. Default is 'nan_euclidean'.
-    keep_empty_features : bool, optional
-        Whether to keep empty features (no known samples). Default is True.
-
-    Returns
-    -------
-    Union[pd.DataFrame, List[pd.DataFrame]]
-        A pandas DataFrame or list of pandas DataFrames with imputed missing values.
-    """
-    imputer = KNNImputer(
-        n_neighbors=n_neighbors,
-        weights=weights,
-        metric=metric,
-        keep_empty_features=keep_empty_features,
-    )
-
-    if isinstance(data, pd.DataFrame):
-        # If it's a single DataFrame, transform it and return immediately
-        return pd.DataFrame(imputer.fit_transform(data), columns=data.columns, index=data.index)
-    else:
-        # Otherwise, use list comprehension to apply the imputer to each DataFrame
-        return [
-            pd.DataFrame(imputer.fit_transform(t), columns=t.columns, index=t.index) for t in data
-        ]
-
-
 def split_df_by_column(df: pd.DataFrame, cov_index_col: str) -> List[pd.DataFrame]:
     """
     Split a DataFrame by unique values of a specified column.
